@@ -1,41 +1,39 @@
 name: flutter_wordpress_app
 description: Flutter news app for wordpress platform
 
-version: 1.0.0+4
+version: 2.0.0+5
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.0
-  url_launcher: ^5.6.0
-  http: ^0.12.2
-  html: ^0.14.0+3
-  flutter_html: ^0.11.0
-  dio: ^3.0.10
+  url_launcher: ^6.0.11
+  http: ^0.13.3
+  html: ^0.15.0
+  flutter_html: ^2.1.2
+  dio: ^4.0.0
   loading: ^1.0.2
-  intl: ^0.16.1
-  share: ^0.6.5+1
-  path_provider: ^1.6.14
-  sqflite: ^1.3.1+1
-  firebase_messaging: ^7.0.0
-  shared_preferences: ^0.5.10
-  flutter_widget_from_html: ^0.5.0+3
-  
+  intl: ^0.17.0
+  share: ^2.0.4
+  path_provider: ^2.0.5
+  sqflite: ^2.0.0+4
+  shared_preferences: ^2.0.8
+  flutter_widget_from_html: ^0.7.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_launcher_icons: "^0.7.3"
+  flutter_launcher_icons: ^0.9.2
 
 flutter_icons:
-  android: "launcher_icon" 
+  android: "launcher_icon"
   ios: true
   image_path: assets/launcher_icon.png
-
 
 flutter:
   # The following line ensures that the Material Icons font is
@@ -61,13 +59,12 @@ flutter:
     - assets/more/contact.png
     - assets/more/share.png
     - assets/more/notification.png
-    
 
   fonts:
     - family: Poppins
       fonts:
         - asset: fonts/Poppins/Poppins-Bold.ttf
-  
+
     - family: Soleil
       fonts:
         - asset: fonts/Soleil/Soleil-Bold.otf
