@@ -7,15 +7,13 @@ import 'package:flutter_wordpress_app/common/constants.dart';
 import 'package:flutter_wordpress_app/models/Comment.dart';
 import 'package:flutter_wordpress_app/widgets/commentBox.dart';
 import 'package:http/http.dart' as http;
-import 'package:loading/indicator/ball_beat_indicator.dart';
-import 'package:loading/loading.dart';
 
 import 'add_comment.dart';
 
 Future<List<dynamic>> fetchComments(int id) async {
   try {
     var response = await http
-        .get("$WORDPRESS_URL/wp-json/wp/v2/comments?post=" + id.toString());
+        .get(Uri.parse("$WORDPRESS_URL/wp-json/wp/v2/comments?post=" + id.toString()));
 
     if (response.statusCode == 200) {
       return json
@@ -33,7 +31,7 @@ Future<List<dynamic>> fetchComments(int id) async {
 class Comments extends StatefulWidget {
   final int commentId;
 
-  Comments(this.commentId, {Key key}) : super(key: key);
+  Comments(this.commentId, {Key? key}) : super(key: key);
   @override
   _CommentsState createState() => _CommentsState();
 }
@@ -88,7 +86,7 @@ Widget commentSection(Future<List<dynamic>> comments) {
     future: comments,
     builder: (context, commentSnapshot) {
       if (commentSnapshot.hasData) {
-        if (commentSnapshot.data.length == 0)
+        if (commentSnapshot.data!.length == 0)
           return Container(
             height: 500,
             alignment: Alignment.center,
@@ -98,7 +96,7 @@ Widget commentSection(Future<List<dynamic>> comments) {
             ),
           );
         return Column(
-            children: commentSnapshot.data.map((item) {
+            children: commentSnapshot.data!.map((item) {
           return InkWell(
             onTap: () {},
             child: commentBox(context, item.author, item.avatar, item.content),
@@ -113,10 +111,6 @@ Widget commentSection(Future<List<dynamic>> comments) {
       return Container(
         alignment: Alignment.center,
         height: 400,
-        child: Loading(
-            indicator: BallBeatIndicator(),
-            size: 60.0,
-            color: Theme.of(context).accentColor),
       );
     },
   );
