@@ -7,8 +7,6 @@ import 'package:flutter_wordpress_app/models/Article.dart';
 import 'package:flutter_wordpress_app/pages/single_Article.dart';
 import 'package:flutter_wordpress_app/widgets/articleBox.dart';
 import 'package:http/http.dart' as http;
-import 'package:loading/indicator/ball_beat_indicator.dart';
-import 'package:loading/loading.dart';
 
 class LocalArticles extends StatefulWidget {
   @override
@@ -17,11 +15,11 @@ class LocalArticles extends StatefulWidget {
 
 class _LocalArticlesState extends State<LocalArticles> {
   List<dynamic> articles = [];
-  Future<List<dynamic>> _futureArticles;
+  Future<List<dynamic>>? _futureArticles;
 
-  ScrollController _controller;
+  ScrollController? _controller;
   int page = 1;
-  bool _infiniteStop;
+  bool _infiniteStop = false;
 
   @override
   void initState() {
@@ -29,20 +27,20 @@ class _LocalArticlesState extends State<LocalArticles> {
     _futureArticles = fetchLocalArticles(1);
     _controller =
         ScrollController(initialScrollOffset: 0.0, keepScrollOffset: true);
-    _controller.addListener(_scrollListener);
+    _controller!.addListener(_scrollListener);
     _infiniteStop = false;
   }
 
   @override
   void dispose() {
     super.dispose();
-    _controller.dispose();
+    _controller!.dispose();
   }
 
   Future<List<dynamic>> fetchLocalArticles(int page) async {
     try {
       http.Response response = await http.get(
-          "$WORDPRESS_URL/wp-json/wp/v2/posts/?categories[]=$PAGE2_CATEGORY_ID&page=$page&per_page=10&_fields=id,date,title,content,custom,link");
+          Uri.parse("$WORDPRESS_URL/wp-json/wp/v2/posts/?categories[]=$PAGE2_CATEGORY_ID&page=$page&per_page=10&_fields=id,date,title,content,custom,link"));
       if (this.mounted) {
         if (response.statusCode == 200) {
           setState(() {
@@ -69,8 +67,8 @@ class _LocalArticlesState extends State<LocalArticles> {
   }
 
   _scrollListener() {
-    var isEnd = _controller.offset >= _controller.position.maxScrollExtent &&
-        !_controller.position.outOfRange;
+    var isEnd = _controller!.offset >= _controller!.position.maxScrollExtent &&
+        !_controller!.position.outOfRange;
     if (isEnd) {
       setState(() {
         page += 1;
@@ -102,7 +100,7 @@ class _LocalArticlesState extends State<LocalArticles> {
             controller: _controller,
             child: Column(
               children: <Widget>[
-                categoryPosts(_futureArticles),
+                categoryPosts(_futureArticles as Future<List<dynamic>>),
               ],
             )),
       ),
@@ -114,11 +112,11 @@ class _LocalArticlesState extends State<LocalArticles> {
       future: futureArticles,
       builder: (context, articleSnapshot) {
         if (articleSnapshot.hasData) {
-          if (articleSnapshot.data.length == 0) return Container();
+          if (articleSnapshot.data!.length == 0) return Container();
           return Column(
             children: <Widget>[
               Column(
-                  children: articleSnapshot.data.map((item) {
+                  children: articleSnapshot.data!.map((item) {
                 final heroId = item.id.toString() + "-latest";
                 return InkWell(
                   onTap: () {
@@ -136,10 +134,7 @@ class _LocalArticlesState extends State<LocalArticles> {
                   ? Container(
                       alignment: Alignment.center,
                       height: 30,
-                      child: Loading(
-                          indicator: BallBeatIndicator(),
-                          size: 60.0,
-                          color: Theme.of(context).accentColor))
+)
                   : Container()
             ],
           );
@@ -150,10 +145,7 @@ class _LocalArticlesState extends State<LocalArticles> {
             alignment: Alignment.center,
             height: 400,
             width: MediaQuery.of(context).size.width - 30,
-            child: Loading(
-                indicator: BallBeatIndicator(),
-                size: 60.0,
-                color: Theme.of(context).accentColor));
+);
       },
     );
   }
