@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_wordpress_app/models/Article.dart';
-import 'package:html/dom.dart' as dom;
 
 Widget articleBox(BuildContext context, Article article, String heroId) {
   return ConstrainedBox(
@@ -21,28 +20,24 @@ Widget articleBox(BuildContext context, Article article, String heroId) {
               child: Column(
                 children: <Widget>[
                   Container(
-                    padding: EdgeInsets.fromLTRB(8, 0, 4, 8),
+                    padding: EdgeInsets.fromLTRB(8, 0, 4, 0),
                     child: Column(
                       children: <Widget>[
                         Container(
                           child: Html(
-                              data: article.title.length > 70
-                                  ? "<h1>" +
-                                      article.title.substring(0, 70) +
-                                      "...</h1>"
-                                  : "<h1>" + article.title + "</h1>",
-                              customTextStyle:
-                                  (dom.Node node, TextStyle baseStyle) {
-                                if (node is dom.Element) {
-                                  switch (node.localName) {
-                                    case "h1":
-                                      return baseStyle.merge(Theme.of(context)
-                                          .textTheme
-                                          .headline1);
-                                  }
-                                }
-                                return baseStyle;
-                              }),
+                              data: article.title!.length > 70
+                                  ? "<h2>" +
+                                      article.title!.substring(0, 70) +
+                                      "...</h2>"
+                                  : "<h2>" + article.title.toString() + "</h2>",
+                              style: {
+                                "h2": Style(
+                                    color: Theme.of(context).primaryColorDark,
+                                    fontWeight: FontWeight.w500,
+                                    fontSize: FontSize.em(1.05),
+                                    padding: EdgeInsets.all(2),
+                                )},
+                          ),
                         ),
                         Container(
                           alignment: Alignment.topLeft,
@@ -57,7 +52,7 @@ Widget articleBox(BuildContext context, Article article, String heroId) {
                                 padding: EdgeInsets.fromLTRB(8, 4, 8, 4),
                                 margin: EdgeInsets.fromLTRB(0, 0, 0, 8),
                                 child: Text(
-                                  article.category,
+                                  article.category.toString(),
                                   style: TextStyle(
                                       color: Colors.black,
                                       fontSize: 11,
@@ -77,7 +72,7 @@ Widget articleBox(BuildContext context, Article article, String heroId) {
                                       width: 4,
                                     ),
                                     Text(
-                                      article.date,
+                                      article.date.toString(),
                                       style:
                                           Theme.of(context).textTheme.caption,
                                     ),
@@ -104,7 +99,7 @@ Widget articleBox(BuildContext context, Article article, String heroId) {
               child: ClipRRect(
                 borderRadius: new BorderRadius.circular(8.0),
                 child: Image.network(
-                  article.image,
+                  article.image.toString(),
                   fit: BoxFit.cover,
                 ),
               ),
@@ -121,7 +116,6 @@ Widget articleBox(BuildContext context, Article article, String heroId) {
                 left: 12,
                 top: 12,
                 child: Card(
-                  color: Theme.of(context).accentColor,
                   child: CircleAvatar(
                     radius: 14,
                     backgroundColor: Colors.transparent,
